[BY-321] Refactor: 위젯 파일 위치 변경 및 코드 구조 수정
