+++ lib/widget/common/delete_report_block_alert.dart
@@ -1,5 +1,5 @@
-import 'package:byourside/widget/fully_rounded_rectangle_button.dart';
-import 'package:byourside/widget/outlined_rounded_rectangle_button.dart';
+import 'package:byourside/widget/common/fully_rounded_rectangle_button.dart';
+import 'package:byourside/widget/common/outlined_rounded_rectangle_button.dart';
 import 'package:flutter/material.dart';
 import 'package:byourside/constants/fonts.dart' as fonts;
 import 'package:byourside/constants/colors.dart' as colors;
@@ -21,7 +21,7 @@ class DeleteReportBlockAlert extends StatelessWidget {
   Widget build(BuildContext context) {
     return AlertDialog(
       shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(30),
+        borderRadius: BorderRadius.circular(10),
       ),
       semanticLabel: message + subMessage,
       content: Column(
@@ -50,12 +50,16 @@ class DeleteReportBlockAlert extends StatelessWidget {
         ],
       ),
       actions: [
-        fullyRoundedRectangleButton('취소하기', () {
-          Navigator.of(context).pop(false); // Return false when canceled
-        }),
-        outlinedRoundedRectangleButton(buttonText, () {
-          onPressed();
-        }),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceAround,
+          children: [
+            fullyRoundedRectangleButton('취소하기', () {
+              Navigator.of(context).pop(false); // Return false when canceled
+            }),
+            outlinedRoundedRectangleButton(buttonText, () {
+              onPressed();
+            }),
+        ])
       ],
     );
   }
