+++ lib/widget/common/title_only_appbar.dart
@@ -1,4 +1,4 @@
-import 'package:byourside/widget/icon_buttons.dart';
+import 'package:byourside/widget/common/icon_buttons.dart';
 import 'package:flutter/material.dart';
 import 'package:byourside/constants/colors.dart' as colors;
 import 'package:byourside/constants/fonts.dart' as fonts;
@@ -14,8 +14,7 @@ PreferredSizeWidget titleOnlyAppbar(BuildContext context, String title, {bool sh
         color: colors.textColor,
         fontSize: 20,
         fontFamily: fonts.font,
-        fontWeight: FontWeight.w700,
-        height: 1.50,
+        fontWeight: FontWeight.w700
       ),
     ),
     leading: showBackButton ? backToPreviousPage(context) : null, // 조건에 따라 뒤로가기 버튼을 표시하거나 숨깁니다.
