+++ lib/widget/community/stream_community_post.dart
@@ -1,6 +1,6 @@
 import 'package:byourside/model/community_post.dart';
 import 'package:byourside/screen/community/community_post_list_tile.dart';
-import 'package:byourside/widget/no_data.dart';
+import 'package:byourside/widget/common/no_data.dart';
 import 'package:flutter/material.dart';
 
 Widget streamCommunityPost(Function loadData) {
